refactor: EventQueue -> NonBlockingEventQueue
