+++ springql-core/src/stream_engine/autonomous_executor/task_executor/generic_worker_pool/generic_worker.rs
@@ -5,7 +5,7 @@ pub(in crate::stream_engine::autonomous_executor) mod generic_worker_thread;
 use std::sync::Arc;
 
 use crate::stream_engine::autonomous_executor::{
-    event_queue::EventQueue,
+    event_queue::non_blocking_event_queue::NonBlockingEventQueue,
     main_job_lock::MainJobLock,
     task_executor::task_worker_thread_handler::TaskWorkerThreadArg,
     worker::worker_handle::{WorkerHandle, WorkerSetupCoordinator, WorkerStopCoordinator},
@@ -22,7 +22,7 @@ pub(super) struct GenericWorker {
 impl GenericWorker {
     pub(super) fn new(
         main_job_lock: Arc<MainJobLock>,
-        event_queue: Arc<EventQueue>,
+        event_queue: Arc<NonBlockingEventQueue>,
         worker_setup_coordinator: Arc<WorkerSetupCoordinator>,
         worker_stop_coordinator: Arc<WorkerStopCoordinator>,
         thread_arg: TaskWorkerThreadArg,
