+++ springql-core/src/stream_engine/autonomous_executor/task_executor/generic_worker_pool/generic_worker/generic_worker_thread.rs
@@ -5,7 +5,7 @@ mod generic_worker_scheduler;
 use std::sync::Arc;
 
 use crate::stream_engine::autonomous_executor::{
-    event_queue::{event::EventTag, EventQueue},
+    event_queue::{event::EventTag, non_blocking_event_queue::NonBlockingEventQueue},
     memory_state_machine::{MemoryState, MemoryStateTransition},
     performance_metrics::{
         metrics_update_command::metrics_update_by_task_execution::MetricsUpdateByTaskExecutionOrPurge,
@@ -46,7 +46,7 @@ impl WorkerThread for GenericWorkerThread {
     fn main_loop_cycle(
         current_state: Self::LoopState, // generic worker's loop cycle does not mutate state (while event handlers do)
         thread_arg: &Self::ThreadArg,
-        event_queue: &EventQueue,
+        event_queue: &NonBlockingEventQueue,
     ) -> Self::LoopState {
         TaskWorkerThreadHandler::main_loop_cycle::<GenericWorkerScheduler>(
             current_state,
@@ -59,7 +59,7 @@ impl WorkerThread for GenericWorkerThread {
         current_state: Self::LoopState,
         pipeline_derivatives: Arc<PipelineDerivatives>,
         thread_arg: &Self::ThreadArg,
-        _event_queue: Arc<EventQueue>,
+        _event_queue: Arc<NonBlockingEventQueue>,
     ) -> Self::LoopState {
         log::debug!(
             "[GenericWorker#{}] got UpdatePipeline event",
@@ -75,7 +75,7 @@ impl WorkerThread for GenericWorkerThread {
         current_state: Self::LoopState,
         metrics: Arc<PerformanceMetrics>,
         thread_arg: &Self::ThreadArg,
-        _event_queue: Arc<EventQueue>,
+        _event_queue: Arc<NonBlockingEventQueue>,
     ) -> Self::LoopState {
         log::debug!(
             "[GenericWorker#{}] got ReplacePerformanceMetrics event",
@@ -91,7 +91,7 @@ impl WorkerThread for GenericWorkerThread {
         current_state: Self::LoopState,
         memory_state_transition: Arc<MemoryStateTransition>,
         thread_arg: &Self::ThreadArg,
-        _event_queue: Arc<EventQueue>,
+        _event_queue: Arc<NonBlockingEventQueue>,
     ) -> Self::LoopState {
         let mut state = current_state;
 
@@ -123,7 +123,7 @@ impl WorkerThread for GenericWorkerThread {
         _current_state: Self::LoopState,
         _metrics: Arc<MetricsUpdateByTaskExecutionOrPurge>,
         _thread_arg: &Self::ThreadArg,
-        _event_queue: Arc<EventQueue>,
+        _event_queue: Arc<NonBlockingEventQueue>,
     ) -> Self::LoopState {
         unreachable!()
     }
@@ -132,7 +132,7 @@ impl WorkerThread for GenericWorkerThread {
         _current_state: Self::LoopState,
         _metrics_summary: Arc<PerformanceMetricsSummary>,
         _thread_arg: &Self::ThreadArg,
-        _event_queue: Arc<EventQueue>,
+        _event_queue: Arc<NonBlockingEventQueue>,
     ) -> Self::LoopState {
         unreachable!()
     }
