+++ springql-core/src/stream_engine/autonomous_executor/worker/worker_handle.rs
@@ -10,7 +10,9 @@ use parking_lot::{Mutex, MutexGuard};
 
 use crate::{
     low_level_rs::SpringConfig,
-    stream_engine::autonomous_executor::{event_queue::EventQueue, main_job_lock::MainJobLock},
+    stream_engine::autonomous_executor::{
+        event_queue::non_blocking_event_queue::NonBlockingEventQueue, main_job_lock::MainJobLock,
+    },
 };
 
 use super::worker_thread::WorkerThread;
@@ -24,7 +26,7 @@ pub(in crate::stream_engine::autonomous_executor) struct WorkerHandle {
 impl WorkerHandle {
     pub(in crate::stream_engine::autonomous_executor) fn new<T: WorkerThread>(
         main_job_lock: Arc<MainJobLock>,
-        event_queue: Arc<EventQueue>,
+        event_queue: Arc<NonBlockingEventQueue>,
         worker_setup_coordinator: Arc<WorkerSetupCoordinator>,
         worker_stop_coordinator: Arc<WorkerStopCoordinator>,
         thread_arg: T::ThreadArg,
